+++ src/dropin-model.js
@@ -34,7 +34,7 @@ var ASYNC_DEPENDENCIES = DEFAULT_PAYMENT_OPTION_PRIORITY.concat(NON_PAYMENT_OPTI
 var DEPENDENCY_READY_CHECK_INTERVAL = 200;
 
 function DropinModel(options) {
-  var vaultedPaymentMethodTypesThatShouldBeHidden = options.merchantConfiguration.vaultedPaymentMethodTypesThatShouldBeHidden;
+  var hiddenVaultedPaymentMethodTypes = options.merchantConfiguration.hiddenVaultedPaymentMethodTypes;
   var self = this;
 
   this.rootNode = options.container;
@@ -48,31 +48,31 @@ function DropinModel(options) {
 
     return total;
   }.bind(this), {});
-  this.vaultedPaymentMethodTypesThatShouldBeHidden =
+  this.hiddenVaultedPaymentMethodTypes =
     constructHiddenPaymentMethodTypes(
-      options.merchantConfiguration.vaultedPaymentMethodTypesThatShouldBeHidden
+      options.merchantConfiguration.hiddenVaultedPaymentMethodTypes
     );
 
-  if (Array.isArray(vaultedPaymentMethodTypesThatShouldBeHidden)) {
-    this.vaultedPaymentMethodTypesThatShouldBeHidden = [];
+  if (Array.isArray(hiddenVaultedPaymentMethodTypes)) {
+    this.hiddenVaultedPaymentMethodTypes = [];
     // Values is not unique (e.g PayPalAccount)
     Object.values(paymentMethodTypes).filter(function (value, index, _self) {
       return _self.indexOf(value) === index;
     }).forEach(
       function (paymentMethod) {
-        if (vaultedPaymentMethodTypesThatShouldBeHidden.indexOf(paymentMethod) !== -1) {
-          self.vaultedPaymentMethodTypesThatShouldBeHidden.push(paymentMethod);
+        if (hiddenVaultedPaymentMethodTypes.indexOf(paymentMethod) !== -1) {
+          self.hiddenVaultedPaymentMethodTypes.push(paymentMethod);
         }
       }
     );
     // Add the default values:
     VAULTED_PAYMENT_METHOD_TYPES_THAT_SHOULD_BE_HIDDEN.forEach(function (paymentMethod) {
-      if (self.vaultedPaymentMethodTypesThatShouldBeHidden.indexOf(paymentMethod) === -1) {
-        self.vaultedPaymentMethodTypesThatShouldBeHidden.push(paymentMethod);
+      if (self.hiddenVaultedPaymentMethodTypes.indexOf(paymentMethod) === -1) {
+        self.hiddenVaultedPaymentMethodTypes.push(paymentMethod);
       }
     });
   } else {
-    this.vaultedPaymentMethodTypesThatShouldBeHidden = [].concat(VAULTED_PAYMENT_METHOD_TYPES_THAT_SHOULD_BE_HIDDEN);
+    this.hiddenVaultedPaymentMethodTypes = [].concat(VAULTED_PAYMENT_METHOD_TYPES_THAT_SHOULD_BE_HIDDEN);
   }
 
   this.failedDependencies = {};
@@ -425,7 +425,7 @@ DropinModel.prototype._getSupportedPaymentMethods = function (paymentMethods) {
 
     if (
       canShowVaultedPaymentMethodType(
-        paymentMethodType, self.vaultedPaymentMethodTypesThatShouldBeHidden)
+        paymentMethodType, self.hiddenVaultedPaymentMethodTypes)
     ) {
       array.push(paymentMethodType);
     }
@@ -502,17 +502,17 @@ function isPaymentOptionEnabled(paymentOption, options) {
 }
 
 function canShowVaultedPaymentMethodType(
-  paymentMethodType, vaultedPaymentMethodTypesThatShouldBeHidden
+  paymentMethodType, hiddenVaultedPaymentMethodTypes
 ) {
   return (
     paymentMethodType &&
-    vaultedPaymentMethodTypesThatShouldBeHidden.indexOf(paymentMethodType) ===
+    hiddenVaultedPaymentMethodTypes.indexOf(paymentMethodType) ===
       -1
   );
 }
 
 function constructHiddenPaymentMethodTypes(paymentMethods) {
-  var vaultedPaymentMethodTypesThatShouldBeHidden = [].concat(
+  var hiddenVaultedPaymentMethodTypes = [].concat(
     VAULTED_PAYMENT_METHOD_TYPES_THAT_SHOULD_ALWAYS_BE_HIDDEN
   );
 
@@ -526,18 +526,18 @@ function constructHiddenPaymentMethodTypes(paymentMethods) {
       }
 
       if (
-        vaultedPaymentMethodTypesThatShouldBeHidden.indexOf(paymentMethodId) >
+        hiddenVaultedPaymentMethodTypes.indexOf(paymentMethodId) >
         -1
       ) {
         // don't add the same payment method type a second time
         return;
       }
 
-      vaultedPaymentMethodTypesThatShouldBeHidden.push(paymentMethodId);
+      hiddenVaultedPaymentMethodTypes.push(paymentMethodId);
     });
   }
 
-  return vaultedPaymentMethodTypesThatShouldBeHidden;
+  return hiddenVaultedPaymentMethodTypes;
 }
 
 module.exports = DropinModel;